alterando o campo da tabela valor

+++ security/database/migrations/2021_02_12_022456_create_registros_table.php
@@ -17,7 +17,7 @@ class CreateRegistrosTable extends Migration
             $table->id();
             $table->string('local');
             $table->string('data');
-            $table->integer('valor');
+            $table->decimal('valor');
             $table->integer('user_id');
             $table->timestamps();
         });
